requirements.txt
python-dotenv>=1.2.0
prompt-toolkit>=3.0.0
boto3>=1.35.0